refactor: Update import paths for load_chunks and remove batch_embedder_checkout script
