+++ scripts/embeddings/batch_embedder_checkout.py
