+++ scripts/embeddings/chunk_embedder.py
@@ -11,7 +11,7 @@ import time
 from scripts.chunking.models import Chunk
 from scripts.core.project_manager import ProjectManager
 from scripts.utils.logger import LoggerManager
-from scripts.utils.chunks_io import load_chunks
+from scripts.utils.chunk_utils import load_chunks
 from scripts.embeddings.embedder_registry import get_embedder
