+++ scripts/embeddings/chunk_embedder_2.py
@@ -10,7 +10,7 @@ import faiss
 from scripts.chunking.models import Chunk
 from scripts.core.project_manager import ProjectManager
 from scripts.utils.logger import LoggerManager
-from scripts.utils.chunks_io import load_chunks
+from scripts.utils.chunk_utils import load_chunks
 from scripts.embeddings.embedder_registry import get_embedder
